Remove itemindent to preserve indentation of subsequent lines
If text wraps to a new line in the same bullet, the text on the new line
appeared *under* the bullet, e.g.,

- This is a long line of text that is going to wrap around soon, but the
new text wraps *under* the bullet.

What I want is for the new text to wrap so there is space under the
bullet, like so:

- This is a long line of text that is going to wrap around soon, but the
  new text wraps *under* the bullet.

Notice the indent in the text in the new line.
By removing itemindent, I am able to accomplish this.